refactor(VolunteerList): разделил на компоненты, прибрался
